+++ src/context/TodosContext.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import { TodosContextType } from './type';
+
+export const TodosContext = createContext<TodosContextType | undefined>(
+  undefined,
+);
